vet_nucleotide_sequence function is working.
I still dont really understand how all of these modificaton to the rna_patter_str and dna_pattern_str work. I will outline them here so I can refer back later. To start off easy, the uppercase and lowercase inside the bracket allow it to accept lowercase characters without throwing an error. the brackets, I believe, allow it to not treat the characters inside as a string. the ^ and $ signify the beginning and end of the pattern, and * allows the pattern to accept any number, including 0, of any and all of the characters.