Finished
I finally finished the part two of this assignment. The lecture on
refular expressions was incredibly helpful. I was still confused on how
to get the patter to accept variables, and my first attempt failed
because it was expecting a string and got a list instead. some googling
lead me to the .join function which made everything work.