Finished
I have the function working, and I think it is done. I had to add
the genetic_code dictionary from the translate.py, which I dont
think should have been necessary. I believe one of the extra
practices that week was to modify the code such that there was no
need to add a genetic_code when calling the function. I did not do
this. Had I done that I believe this would have been much simplier.